trying to fix crashes on window minimize

+++ src/ast/type.cpp
@@ -293,7 +293,8 @@ const TypeImpl *TypeRegistry::_decode(luisa::string_view desc) noexcept {
                 info->dimension);
         }
         info->alignment = std::min(
-            elem->size() * (info->dimension == 3 ? 4 : info->dimension), 16ull);
+            elem->size() * (info->dimension == 3 ? 4 : info->dimension),
+            static_cast<size_t>(16u));
         info->size = luisa::align(elem->size() * info->dimension, info->alignment);
     } else if (type_identifier == "matrix"sv) {
         info->tag = Type::Tag::MATRIX;
